plugins/jokeapi/requirements.txt
qrcode
fonts
font-roboto
Pillow